refactor: fixText extracted for testing
